formatter: ruby: Successfully read output from Go executable
